# openapi.yaml
openapi: 3.1.0
info:
  title: search
  version: "1.0.0"
  description: >-
    Full-text search over your own data: named indexes, bulk-indexed documents, BM25-ranked
    queries with field weights, attribute filters and facet counts, highlighted results, prefix
    suggestions. SQLite FTS5, no external search engine. Errors are always
    `{ "error": { "code", "message", "details?" } }`.
servers:
  - url: http://localhost:3010
    description: Local default (PORT=3010).
security:
  - BearerAuth: []
...
paths:
...
  /v1/indexes:
    post:
      operationId: search.indexes.create
      tags: [indexes]
      summary: Create an index
      description: >-
        Requires a `write`-role key. If the key is index-scoped, the requested `name` must be in
        its scope (checked after body validation, before creation) or this returns 403 — a scoped
        key can preemptively be issued for an index that does not exist yet.
      x-stack-visibility: authenticated
      x-required-role: write
      x-index-scoped: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIndexRequest"
      responses:
        "201":
          description: Index created.
          headers:
            location:
              description: "`/v1/indexes/{name}` of the new index."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexResponse"
        "400":
          description: >-
            `INVALID_QUERY` (a weight is out of the 0-100 range, or a facet key is not a valid
            attribute key, or more than 20 facets) or `VALIDATION_FAILED` (request body fails
            schema validation, e.g. bad `name` pattern).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/IndexExists"
        "429":
          $ref: "#/components/responses/RateLimited"

    get:
      operationId: search.indexes.list
      tags: [indexes]
      summary: List visible indexes
      description: >-
        Requires a `read`-role key. Returns only the indexes in the key's scope (or every index
        for an unscoped key), each with its live document count and search counter.
      x-stack-visibility: authenticated
      x-required-role: read
      x-index-scoped: true
      responses:
        "200":
          description: Visible indexes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/indexes/{name}:
    parameters:
      - $ref: "#/components/parameters/IndexName"
    get:
      operationId: search.indexes.get
      tags: [indexes]
      summary: Get one index
      description: Requires a `read`-role key scoped to this index (or unscoped).
      x-stack-visibility: authenticated
      x-required-role: read
      x-index-scoped: true
      responses:
        "200":
          description: The index.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/IndexNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"

    patch:
      operationId: search.indexes.patch
      tags: [indexes]
      summary: Update an index
      description: >-
        Requires a `write`-role key scoped to this index (or unscoped). At least one of
        `description`, `weights`, `facets` is required; each given field replaces the
        corresponding stored value in full (`weights`/`facets` are not merged key-by-key beyond
        what the domain layer does internally with existing weights).
      x-stack-visibility: authenticated
      x-required-role: write
      x-index-scoped: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchIndexRequest"
      responses:
        "200":
          description: The updated index.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexResponse"
        "400":
          description: >-
            `INVALID_QUERY` (a weight out of range, or an invalid/too-many facet keys) or
            `VALIDATION_FAILED` (body fails schema validation, e.g. empty patch object).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/IndexNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"

    delete:
      operationId: search.indexes.delete
      tags: [indexes]
      summary: Delete an index
      description: >-
        Requires a `write`-role key scoped to this index (or unscoped). Deletes the index and
        every document in it, in one transaction.
      x-stack-visibility: authenticated
      x-required-role: write
      x-index-scoped: true
      responses:
        "204":
          description: Index (and its documents) deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/IndexNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/indexes/{name}/clear:
    parameters:
      - $ref: "#/components/parameters/IndexName"
    post:
      operationId: search.indexes.clear
      tags: [indexes]
      summary: Remove every document from an index
      description: >-
        Requires a `write`-role key scoped to this index (or unscoped). The index itself (its
        settings) is kept; only its documents are removed.
      x-stack-visibility: authenticated
      x-required-role: write
      x-index-scoped: true
      responses:
        "200":
          description: Documents removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClearResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/IndexNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/indexes/{name}/documents:
    parameters:
      - $ref: "#/components/parameters/IndexName"
    put:
      operationId: search.documents.upsert
      tags: [documents]
      summary: Bulk upsert documents
      description: >-
        Requires a `write`-role key scoped to this index (or unscoped). Up to `MAX_BATCH`
        (default 500) documents per request, validated and indexed in one transaction: either
        every document lands or none do. Duplicate `id`s within the same batch are rejected.
      x-stack-visibility: authenticated
      x-required-role: write
      x-index-scoped: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertRequest"
      responses:
        "200":
          description: Counts of created vs. updated documents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpsertResult"
        "400":
          description: >-
            `INVALID_DOCUMENT` (a document fails domain validation: missing/empty `title`,
            disallowed `attrs` key, duplicate `id`s in the batch, ...) or `VALIDATION_FAILED`
            (body fails schema validation).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/IndexNotFound"
        "413":
          description: >-
            `BATCH_TOO_LARGE` (more than `MAX_BATCH` documents in the request) or
            `DOCUMENT_TOO_LARGE` (one document's serialized size exceeds `MAX_DOC_BYTES`, default
            65536 bytes).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"

    get:
      operationId: search.documents.browse
      tags: [documents]
      summary: Browse documents (no query)
      description: >-
        Requires a `read`-role key scoped to this index (or unscoped). Equivalent to an empty
        search sorted newest-first; every returned item omits `score` entirely (it is stripped,
        not `null`, unlike a real search hit). Default `limit` is **50** here (search's default is
        20), clamped to `maxPage` (default 100, `MAX_PAGE`). Default `offset` is 0.
      x-stack-visibility: authenticated
      x-required-role: read
      x-index-scoped: true
      parameters:
        - name: limit
          in: query
          description: Page size. Default 50, clamped to `maxPage` (default 100).
          schema:
            $ref: "#/components/schemas/LimitParam"
        - name: offset
          in: query
          description: >-
            Row offset. Default 0. Rejected with 400 `INVALID_QUERY` if greater than `maxOffset`
            (default 10000, `MAX_OFFSET`).
          schema:
            $ref: "#/components/schemas/OffsetParam"
      responses:
        "200":
          description: A page of documents, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrowseResponse"
        "400":
          description: >-
            `INVALID_QUERY` (`offset` greater than `maxOffset`) or `VALIDATION_FAILED` (`limit`/
            `offset` fail their pattern, or an unknown query parameter is present — this endpoint's
            querystring schema is strict, unlike the search endpoints).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/IndexNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/indexes/{name}/documents/{id}:
    parameters:
      - $ref: "#/components/parameters/IndexName"
      - $ref: "#/components/parameters/DocId"
    get:
      operationId: search.documents.get
      tags: [documents]
      summary: Get one document
      description: Requires a `read`-role key scoped to this index (or unscoped).
      x-stack-visibility: authenticated
      x-required-role: read
      x-index-scoped: true
      responses:
        "200":
          description: The document.
          content:
            application/json:
              schema:
                type: object
                required: [document]
                properties:
                  document:
                    $ref: "#/components/schemas/DocumentView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "`INDEX_NOT_FOUND` or `DOCUMENT_NOT_FOUND`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"

    delete:
      operationId: search.documents.delete
      tags: [documents]
      summary: Delete one document
      description: Requires a `write`-role key scoped to this index (or unscoped).
      x-stack-visibility: authenticated
      x-required-role: write
      x-index-scoped: true
      responses:
        "204":
          description: Document deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "`INDEX_NOT_FOUND` or `DOCUMENT_NOT_FOUND`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/indexes/{name}/search:
    parameters:
      - $ref: "#/components/parameters/IndexName"
    get:
      operationId: search.query.get
      tags: [search]
      summary: Search (query-string form)
      description: >-
        Requires a `read`-role key scoped to this index (or unscoped). Query form of the same
        search as `POST` on this path, sharing the same response shape (`SearchResult`) — but the
        two are intentionally **not** validated the same way:

        This endpoint's querystring schema is permissive: any additional query parameter is
        accepted as long as its value is a string of at most 2000 characters
        (`additionalProperties: { type: string, maxLength: 2000 }`), regardless of its name. Of
        those, only parameters whose name starts with the literal prefix `filter.` are actually
        consumed — `filter.<attrKey>=v1,v2` becomes an AND-across-keys, OR-within-key attribute
        filter (comma-split, one attribute key per parameter). Any other unrecognized parameter is
        accepted by validation and then silently ignored. This dynamic `filter.*`-prefix pattern
        cannot be expressed as a fixed OpenAPI parameter list; `filter.<key>` here documents the
        pattern, not a literal parameter named `<key>`.

        The `POST` form below is the strict opposite: `filters` is one explicit object field and
        the request body schema rejects any unknown field (`additionalProperties: false`).

        Default `limit` is **20** here (browse's default is 50), clamped to `maxPage` (default
        100). Default `offset` is 0, `sort` is `relevance`, `highlight` is `false`.
      x-stack-visibility: authenticated
      x-required-role: read
      x-index-scoped: true
      parameters:
        - name: q
          in: query
          description: >-
            Search text: words are AND'd, `"phrases"` match as a unit, `-word` excludes, the last
            word matches as a prefix. Empty or omitted means "browse" (no ranking, `score` is
            `null` on every hit).
          schema:
            type: string
            maxLength: 500
        - name: limit
          in: query
          description: Page size. Default 20, clamped to `maxPage` (default 100).
          schema:
            $ref: "#/components/schemas/LimitParam"
        - name: offset
          in: query
          description: >-
            Row offset. Default 0. Rejected with 400 `INVALID_QUERY` if greater than `maxOffset`
            (default 10000).
          schema:
            $ref: "#/components/schemas/OffsetParam"
        - name: highlight
          in: query
          description: >-
            `"true"` to include `highlights` on every hit. Note: a string enum here
            (`"true"`/`"false"`), not a real boolean — this is the querystring form; the `POST`
            body form below uses a real JSON boolean for the same field.
          schema:
            type: string
            enum: ["true", "false"]
        - name: facets
          in: query
          description: >-
            Comma-separated attribute keys to compute facet counts for (up to 20). Each key must
            be a valid attribute key or `tags`, or this returns 400 `INVALID_QUERY`.
          schema:
            type: string
            maxLength: 1000
        - name: sort
          in: query
          schema:
            $ref: "#/components/schemas/SortParam"
        - name: filter.<key>
          in: query
          description: >-
            Dynamic pattern, not a literal parameter name — see the operation description. One
            occurrence per filtered attribute key, e.g. `filter.color=red,blue&filter.brand=Mavi`.
          required: false
          schema:
            type: string
            maxLength: 2000
      responses:
        "200":
          description: Ranked (or browsed) results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"
        "400":
          description: >-
            `INVALID_QUERY` (offset beyond `maxOffset`, more than 32 search terms, an invalid
            `facets` key, or an invalid `filter.<key>` attribute key / more than 50 values for one
            filter key).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/IndexNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"

    post:
      operationId: search.query.post
      tags: [search]
      summary: Search (JSON body form)
      description: >-
        Requires a `read`-role key scoped to this index (or unscoped). Body form of the same
        search as `GET` on this path, sharing the same response shape (`SearchResult`). Unlike the
        `GET` form, the request body schema is strict: `additionalProperties: false`, `filters` is
        one explicit object field (`{ "<attrKey>": ["v1", "v2"] }`, up to 20 keys, up to 50 values
        per key), and `highlight` is a real JSON boolean. Default `limit` is 20, clamped to
        `maxPage` (default 100); default `offset` is 0; default `sort` is `relevance`.
      x-stack-visibility: authenticated
      x-required-role: read
      x-index-scoped: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequestBody"
      responses:
        "200":
          description: Ranked (or browsed) results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"
        "400":
          description: >-
            `INVALID_QUERY` (offset beyond `maxOffset`, more than 32 search terms, an invalid
            `facets`/`filters` attribute key) or `VALIDATION_FAILED` (body fails schema
            validation, e.g. an unknown field).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/IndexNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/indexes/{name}/suggest:
    parameters:
      - $ref: "#/components/parameters/IndexName"
    get:
      operationId: search.suggest
      tags: [search]
      summary: Title prefix suggestions
      description: >-
        Requires a `read`-role key scoped to this index (or unscoped). Type-ahead on titles only:
        folds and tokenizes `q`, matches titles whose words start with the typed words. Response
        items carry **only** `id` and `title` — this is not the full `DocumentView` returned by
        search/browse/get.
      x-stack-visibility: authenticated
      x-required-role: read
      x-index-scoped: true
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: limit
          in: query
          description: Default 10, clamped to `maxPage` (default 100).
          schema:
            $ref: "#/components/schemas/LimitParam"
      responses:
        "200":
          description: Up to `limit` title matches.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/IndexNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/stats:
    get:
      operationId: search.stats.get
      tags: [stats]
      summary: Aggregate stats
      description: >-
        Requires a `read`-role key. Not index-path-scoped (there is no `:name` parameter), but the
        aggregation itself only includes indexes visible to the key's own scope, unlike
        `GET /metrics` which reports every index regardless of scope.
      x-stack-visibility: authenticated
      x-required-role: read
      x-index-scoped: true
      responses:
        "200":
          description: Aggregate counters across visible indexes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >-
        `Authorization: Bearer <secret>`. Keys are configured as `id:secret[:role[:indexes]]`;
        `role` is `read`, `write` or `readwrite` (default), and `indexes` is an optional
        `+`-joined allowlist of index names (omitted = every index). A route whose path names an
        index (`:name`) is scope-gated: a key with a non-null index list may only touch indexes in
        that list, and gets 403 `FORBIDDEN` for any other index, including one that does not exist
        yet. Missing/invalid credentials return 401 `UNAUTHORIZED` with a `www-authenticate:
        Bearer` header; an insufficient role or an out-of-scope index both return 403 `FORBIDDEN`.
...
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: >-
        `FORBIDDEN` — the API key's role does not permit this operation, or the key is scoped to a
        different set of indexes and this index is not in it.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: >-
        `RATE_LIMITED` — more than `RATE_LIMIT_MAX` (default 1200) requests in the trailing
        minute for this API key. Only applies to `/v1/*` routes.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    IndexExists:
      description: "`INDEX_EXISTS` — an index with this name already exists."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    IndexResponse:
      type: object
      required: [index]
      properties:
        index:
          $ref: "#/components/schemas/IndexView"
...
    CreateIndexRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          $ref: "#/components/schemas/IndexNameValue"
        description:
          type: string
          maxLength: 500
        weights:
          $ref: "#/components/schemas/WeightsInput"
        facets:
          $ref: "#/components/schemas/FacetKeys"
...